Update app from Google Podcasts to Pocket Casts
The changes made are to reflect the switch from Google Podcasts to Pocket Casts. These alterations can be seen in the README file, including the motivation section, new releases section and the flowchart of the application's sequence.

Pocket Casts is now the preferred source following the turndown of Google Podcasts in 2024. Changes in source URLs are reflected in the flowcharts and text explanations.